// Assets-Uncapped/front-portal-develop/src/services/api/agreements/models/DetailedAgreementDTO.ts
export const DetailedAgreementDTODetailedProductTypeEnum = {
    Rbf: 'RBF',
    Fixed: 'FIXED',
    InterestRate: 'INTEREST_RATE',
    LineOfCreditFixedFrame: 'LINE_OF_CREDIT_FIXED_FRAME',
    LineOfCreditFixedDraw: 'LINE_OF_CREDIT_FIXED_DRAW',
    LineOfCreditInterestRate: 'LINE_OF_CREDIT_INTEREST_RATE',
    DailyPayout: 'DAILY_PAYOUT',
    Unknown: 'UNKNOWN'
} as const;
export type DetailedAgreementDTODetailedProductTypeEnum = typeof DetailedAgreementDTODetailedProductTypeEnum[keyof typeof DetailedAgreementDTODetailedProductTypeEnum];
// ...
export const DetailedAgreementDTODebtProviderEnum = {
    Uncapped: 'UNCAPPED',
    Fortress: 'FORTRESS',
    InitialPortfolioReceivables: 'INITIAL_PORTFOLIO_RECEIVABLES',
    InitialPortfolioReceivablesAbs: 'INITIAL_PORTFOLIO_RECEIVABLES_ABS',
    UncappedLtd: 'UNCAPPED_LTD',
    UncappedFinance1: 'UNCAPPED_FINANCE_1',
    UncappedFinance2: 'UNCAPPED_FINANCE_2',
    UncappedFinance3: 'UNCAPPED_FINANCE_3',
    UncappedTechGermany: 'UNCAPPED_TECH_GERMANY',
    UncappedManagementGmbh: 'UNCAPPED_MANAGEMENT_GMBH',
    Marcus: 'MARCUS',
    Fasanara: 'FASANARA',
    Unknown: 'UNKNOWN'
} as const;
export type DetailedAgreementDTODebtProviderEnum = typeof DetailedAgreementDTODebtProviderEnum[keyof typeof DetailedAgreementDTODebtProviderEnum];